+++ src/activitypub/handlers/create.rs
@@ -174,7 +174,7 @@ pub async fn get_object_attachments(
             };
             let attachment_url = attachment.url
                 .ok_or(ValidationError("attachment URL is missing"))?;
-            let (file_name, file_size, maybe_media_type) = match fetch_file(
+            let (file_name, file_size, media_type) = match fetch_file(
                 instance,
                 &attachment_url,
                 attachment.media_type.as_deref(),
@@ -193,20 +193,20 @@ pub async fn get_object_attachments(
                 },
             };
             log::info!("downloaded attachment {}", attachment_url);
-            downloaded.push((file_name, file_size, maybe_media_type));
+            downloaded.push((file_name, file_size, media_type));
             // Stop downloading if limit is reached
             if downloaded.len() >= ATTACHMENT_LIMIT {
                 log::warn!("too many attachments");
                 break;
             };
         };
-        for (file_name, file_size, maybe_media_type) in downloaded {
+        for (file_name, file_size, media_type) in downloaded {
             let db_attachment = create_attachment(
                 db_client,
                 &author.id,
                 file_name,
                 file_size,
-                maybe_media_type,
+                media_type,
             ).await?;
             attachments.push(db_attachment.id);
         };
@@ -289,7 +289,7 @@ pub async fn handle_emoji(
         Err(DatabaseError::NotFound("emoji")) => None,
         Err(other_error) => return Err(other_error.into()),
     };
-    let (file_name, file_size, maybe_media_type) = match fetch_file(
+    let (file_name, file_size, media_type) = match fetch_file(
         instance,
         &tag.icon.url,
         tag.icon.media_type.as_deref(),
@@ -302,17 +302,12 @@ pub async fn handle_emoji(
             return Ok(None);
         },
     };
-    let media_type = match maybe_media_type {
-        Some(media_type) if EMOJI_MEDIA_TYPES.contains(&media_type.as_str()) => {
-            media_type
-        },
-        _ => {
-            log::warn!(
-                "unexpected emoji media type: {:?}",
-                maybe_media_type,
-            );
-            return Ok(None);
-        },
+    if !EMOJI_MEDIA_TYPES.contains(&media_type.as_str()) {
+        log::warn!(
+            "unexpected emoji media type: {}",
+            media_type,
+        );
+        return Ok(None);
     };
     log::info!("downloaded emoji {}", tag.icon.url);
     let image = EmojiImage { file_name, file_size, media_type };